+++ src/pages/LoginPage.tsx
@@ -1,10 +1,10 @@
-import AuthPageLayout from '../components/AuthPageLayout';
-import LoginForm from '../components/Forms/LoginForm';
+import AuthPageLayout from "../components/AuthPageLayout";
+import { LoginFormContainer } from "../components/Forms/LoginForm/LoginFormContainer";
 
 export default function LoginPage() {
   return (
     <AuthPageLayout imageURL="images/img_login.png">
-      <LoginForm />
+      <LoginFormContainer />
     </AuthPageLayout>
   );
 }
